ニックネームにユーザーネームそのまま入れた。本当はいらないかもしれないが、とりあえずニックネームカラム残しとく。indexに集約してるから名前表示にニックネームあった方が便利は便利

+++ database/migrations/2020_06_03_175714_create_user_infos_table.php
@@ -15,7 +15,7 @@ class CreateUserInfosTable extends Migration
     {
         Schema::create('user_infos', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->Integer('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->string('nickname', 20);
             $table->boolean('gender');
             $table->tinyInteger('age');
